backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.9.2
joblib==1.2.0
pandas==2.1.4
scikit-learn==1.3.2
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.26.4
python-multipart==0.0.6
websockets==12.0 